consolidate models into single file
