+++ pyomo/contrib/incidence_analysis/tests/test_interface.py
@@ -23,6 +23,10 @@ from pyomo.contrib.incidence_analysis.triangularize import block_triangularize
 from pyomo.contrib.incidence_analysis.dulmage_mendelsohn import (
         dulmage_mendelsohn,
         )
+from pyomo.contrib.incidence_analysis.tests.models_for_testing import (
+        make_gas_expansion_model,
+        make_degenerate_solid_phase_model,
+        )
 if scipy_available:
     from pyomo.contrib.pynumero.interfaces.pyomo_nlp import PyomoNLP
 if networkx_available:
@@ -30,87 +34,6 @@ if networkx_available:
 from pyomo.contrib.pynumero.asl import AmplInterface
 
 import pyomo.common.unittest as unittest
-
-def make_gas_expansion_model(N=2):
-    """
-    This is the simplest model I could think of that has a
-    subsystem with a non-trivial block triangularization.
-    Something like a gas (somehow) undergoing a series
-    of isentropic expansions.
-    """
-    m = pyo.ConcreteModel()
-    m.streams = pyo.Set(initialize=range(N+1))
-    m.rho = pyo.Var(m.streams, initialize=1)
-    m.P = pyo.Var(m.streams, initialize=1)
-    m.F = pyo.Var(m.streams, initialize=1)
-    m.T = pyo.Var(m.streams, initialize=1)
-
-    m.R = pyo.Param(initialize=8.31)
-    m.Q = pyo.Param(m.streams, initialize=1)
-    m.gamma = pyo.Param(initialize=1.4*m.R.value)
-
-    def mbal(m, i):
-        if i == 0:
-            return pyo.Constraint.Skip
-        else:
-            return m.rho[i-1]*m.F[i-1] - m.rho[i]*m.F[i] == 0
-    m.mbal = pyo.Constraint(m.streams, rule=mbal)
-
-    def ebal(m, i):
-        if i == 0:
-            return pyo.Constraint.Skip
-        else:
-            return (
-                    m.rho[i-1]*m.F[i-1]*m.T[i-1] +
-                    m.Q[i] -
-                    m.rho[i]*m.F[i]*m.T[i] == 0
-                    )
-    m.ebal = pyo.Constraint(m.streams, rule=ebal)
-
-    def expansion(m, i):
-        if i == 0:
-            return pyo.Constraint.Skip
-        else:
-            return m.P[i]/m.P[i-1] - (m.rho[i]/m.rho[i-1])**m.gamma == 0
-    m.expansion = pyo.Constraint(m.streams, rule=expansion)
-
-    def ideal_gas(m, i):
-        return m.P[i] - m.rho[i]*m.R*m.T[i] == 0
-    m.ideal_gas = pyo.Constraint(m.streams, rule=ideal_gas)
-
-    return m
-
-
-def make_degenerate_solid_phase_model():
-    """
-    From the solid phase thermo package of a moving bed chemical looping
-    combustion reactor. This example was first presented in [1]
-    
-    [1] Parker, R. Nonlinear programming strategies for dynamic models of
-    chemical looping combustion reactors. Pres. AIChE Annual Meeting, 2020.
-
-    """
-    m = pyo.ConcreteModel()
-    m.components = pyo.Set(initialize=[1,2,3])
-    m.x = pyo.Var(m.components, initialize=1/3)
-    m.flow_comp = pyo.Var(m.components, initialize=10)
-    m.flow = pyo.Var(initialize=30)
-    m.rho = pyo.Var(initialize=1)
-
-    # These are rough approximations of the relevant equations, with the same
-    # incidence.
-    m.sum_eqn = pyo.Constraint(expr=sum(m.x[j] for j in m.components) - 1 == 0)
-    m.holdup_eqn = pyo.Constraint(m.components, expr={
-        j: m.x[j]*m.rho - 1 == 0 for j in m.components
-        })
-    m.density_eqn = pyo.Constraint(expr=
-            1/m.rho - sum(1/m.x[j] for j in m.components) == 0
-            )
-    m.flow_eqn = pyo.Constraint(m.components, expr={
-        j: m.x[j]*m.flow - m.flow_comp[j] == 0 for j in m.components
-        })
-
-    return m
 
 
 @unittest.skipUnless(networkx_available, "networkx is not available.")
